fix: route

+++ Frontend/src/Users.tsx
@@ -172,7 +172,7 @@ const Users = () => {
   };
 
   const handelDelete = async (userId: string) => {
-    
+    if (!confirm("Are you sure you want to delete this Prodct?")) return;
     try {
       const res = await fetch(`${API}/auth/deleteUsers`, {
         method: "DELETE",
